整理钱包相关接口

+++ soho-extend/soho-wallet/soho-wallet-api/src/main/java/work/soho/wallet/api/service/WalletUserApiService.java
@@ -1,5 +1,5 @@
 package work.soho.wallet.api.service;
 
 public interface WalletUserApiService {
-    Boolean verificationPayPassword(Integer userId, String payPassword);
+    Boolean verificationPayPassword(Long userId, String payPassword);
 }
